feat(cart): enhance cart functionality with backend integration
- Added new endpoints for cart operations including adding, updating, and removing items.
- Integrated cart service with backend API for authenticated users.
- Improved error handling and user feedback during cart operations.
- Updated CartController to manage user-specific cart items.
- Enhanced frontend cart component to reflect changes in real-time and handle user authentication.
- Added logging for better debugging and tracking of cart actions.

+++ e-commerce-backend/src/main/java/com/example/repositories/CartItemRepository.java
@@ -1,13 +1,13 @@
 package com.example.repositories;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.example.models.CartItem;
 
-import java.util.List;
-
 @Repository
 public interface CartItemRepository extends JpaRepository<CartItem, Long> {
-    List<CartItem> findByUser_Id(Long userId);
+    List<CartItem> findByUser_Id(Integer userId);
 }
